Fix performance and dead code issues in interactive components
- Remove unused ScrollDrivenScene component and export
- Add IntersectionObserver to BrainMap3D to gate scroll listener
- Prevents unnecessary scroll events when component is offscreen
- Prevents Canvas mount/unmount thrashing near viewport boundary
