+++ src/components/interactive/index.ts
@@ -1,2 +1,1 @@
 export { DistractionSimulator } from './DistractionSimulator'
-export { ScrollDrivenScene } from './ScrollDrivenScene'
